extract audit assertions and use them to update more tests
